Added mobs prefabs with animations, updated all player's weapons and mobs attack speed for better gameplay, adjusted gold drops of mobs and adjusdted xp table, fixed minor damage calculation mistake accounting damage dealt after reduction of armor points and element debuff

+++ Assets/Scripts/Damage.cs
@@ -7,7 +7,7 @@ public struct Damage
     public bool isHaveWeapon;
     public bool isMelee;
     public Vector3 origin;
-    public int damagePoints;
+    public float damagePoints;
     public float knockbackForce;
     public int attackRange;
     public float cooldown;
